requirements.txt
timm==0.4.12
Pillow
blobfile
mypy
numpy==1.26.4
pytest
requests
einops
#deepspeed==0.4.0
scipy
pyhealth==1.1.4
h5py
mne==1.4.2
rich
torchmetrics